# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Подсистема 'Каршеринг'",
        "contact": {
            "email": "carsh-app@example.com"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://192.168.0.35:80",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/cars": {
            "get": {
                "tags": [
                    "Машины"
                ],
                "summary": "Вернуть полный список с информацией по машинам",
                "description": "Display a listing of the resource.",
                "operationId": "/api/cars(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает полный список с информацией по машинам"
                    }
                }
            },
            "post": {
                "tags": [
                    "Машины"
                ],
                "summary": "Добавить новое поле с информацией по машине",
                "description": "Store a newly created resource in storage.",
                "operationId": "/api/cars(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Добавляет новое поле с информацией по машине"
                    }
                }
            }
        },
...
        "/api/cars/22": {
            "put": {
                "tags": [
                    "Машины"
                ],
                "summary": "Редактировать поле с информацией по конкретной машине",
                "description": "Update the specified resource in storage.",
                "operationId": "/api/cars/{cars}(PUT)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирует поле с информацией по конкретной машине"
                    }
                }
            }
        },
...
        "/api/dvstype": {
            "get": {
                "tags": [
                    "Типы бензина"
                ],
                "summary": "Получение списка типов бензина",
                "description": "Display a listing of the resource.",
                "operationId": "/api/dvstype(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список типов бензина"
                    }
                }
            },
            "post": {
                "tags": [
                    "Типы бензина"
                ],
                "summary": "Добавить новый тип бензина",
                "description": "Store a newly created resource in storage.",
                "operationId": "/api/dvstype(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsDvsTypesStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Добавляет новый тип бензина"
                    }
                }
            }
        },
...
        "/api/dvstype/5": {
            "put": {
                "tags": [
                    "Типы бензина"
                ],
                "summary": "Редактировать поле с информацией по конкретному типу бензина",
                "description": "Update the specified resource in storage.",
                "operationId": "/api/dvstype/{dvstype}(PUT)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsDvsTypesStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирует поле с информацией по конкретному типу бензина"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Типы бензина"
                ],
                "summary": "Удалить поле с информацией по конкретному типу бензина",
                "description": "Remove the specified resource from storage.",
                "operationId": "/api/dvstype/{dvstype}(DELETE)",
                "responses": {
                    "200": {
                        "description": "Удаляет поле с информацией по конкретному типу бензина"
                    }
                }
            }
        },
        "/api/carsmarks": {
            "get": {
                "tags": [
                    "Марки машин"
                ],
                "summary": "Получение списка марок машин",
                "description": "Display a listing of the resource.",
                "operationId": "/api/carsmarks(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список марок машин"
                    }
                }
            },
            "post": {
                "tags": [
                    "Марки машин"
                ],
                "summary": "Добавить новую марку машины",
                "description": "Store a newly created resource in storage.",
                "operationId": "/api/carsmarks(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsMarksStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Добавляет новую марку машины"
                    }
                }
            }
        },
...
        "/api/carsmarks/5": {
            "put": {
                "tags": [
                    "Марки машин"
                ],
                "summary": "Редактировать поле с информацией по конкретной марки машины",
                "description": "Update the specified resource in storage.",
                "operationId": "/api/carsmarks/{carsmarks}(PUT)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsMarksStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирует поле с информацией по конкретной марки машины"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Марки машин"
                ],
                "summary": "Удалить поле с информацией по конкретной марки машины",
                "description": "Remove the specified resource from storage.",
                "operationId": "/api/carsmarks/{carsmarks}(DELETE)",
                "responses": {
                    "200": {
                        "description": "Удаляет поле с информацией по конкретной марки машины"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CarsDvsTypesStoreRequest": {
                "required": [
                    "car_dvs_types"
                ],
                "properties": {
                    "dvs_type": {
                        "description": "тестовый тип двигателя машины",
                        "type": "string",
                        "example": "гидрогелевый"
                    }
                },
                "type": "object"
            },
            "CarsMarksStoreRequest": {
                "required": [
                    "cars_marks"
                ],
                "properties": {
                    "car_mark": {
                        "description": "марка машины",
                        "type": "string",
                        "example": "mazda"
                    }
                },
                "type": "object"
            },
...
            "CarsStoreRequest": {
                "required": [
                    "cars"
                ],
                "properties": {
                    "cars_models_id": {
                        "description": "id модели машины",
                        "type": "integer",
                        "example": "1"
                    },
                    "car_number": {
                        "description": "номер машины",
                        "type": "string",
                        "example": "к123не"
                    },
                    "car_year": {
                        "description": "год выпуска машины",
                        "type": "integer",
                        "example": "2019"
                    },
                    "cars_dvs_type_id": {
                        "description": "тип двигателя",
                        "type": "integer",
                        "example": "3"
                    },
                    "cars_status_id": {
                        "description": "id ",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        }
    }
}